Picasso Image Library Added
Added Picasso image library to the project libraries. This will make
it easy to load images from the assets folder in code into with a single
line of code (See InventoryMenu.java for an example).

Removed the bitmap_util classes, as those are no longer needed for this
project. The Picasso library should be much simpler.