Added a splash screen and player create screen.
Also added some details (name, class, image) to the Player class,
and changed instantiation somewhat - a Singleton class might not
be optimal in the long run - but we shouldn't pretend that we
don't have a singleton when we really do.

And bugs: Changing the player's avatar is buggy, and starting
MainActivity currently causes the game to crash. I'll take a look
at these next.