An item overhaul.
There are now three major kinds of items:
Equipment, Consumables, and Boondogles.
(For reference, a boondogle is a waste of money)
All of them are inherited from the Item class and with luck,
should be able to do everything they need to being accessed
through that. I also cleaned up some unused imports while I was
at it.