Demo Version Commit
Notes:

1. You buy the broomstick item from the shop first. It isn't automatically
added. It is free.
2. Go to the dungeon page. There's a treasure room, monster room and
stairs soom in a 2x2 grid. You can kill the monster in one hit and get the
treasure.
3. When you go to the second floor, the monster takes 2 hits to kill and
kills you in a single hit. This is where you die.
4. Be sure to make a quest that's at least Medium difficulty, or it won't
give enough gold to buy the revival potion.
5. When you return to the dungeon, there is a bug - you are still in the
same room as the monster (this needs to be fixed in the general app
itself). Just click on the same monster room you're in, and you should
fight it again.
6. When you fight the monster again it will die in a single hit.